requirements.txt
alembic==1.7.6
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.6
MarkupSafe==1.1.1
Pillow==9.0.1
psycopg2==2.8.6
pycparser==2.20
six==1.15.0
SQLAlchemy==1.4.7
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.7.0
